Explicitly get all traces on input.
Prior, Digitizer took all untagged traces.  These two should be equivalent in
the context of simulation.

In the, until now nonexistent, case of reading a voltage frame from a FrameFileSink-produced file, the '*' bets (mis)interpreted to be a trace tag instead of being a wild card and this caused Digitizer to not get any traces.

The entire situation with frame and trace tags is a big jumbled and deserves to be straightened out globally.  But, for now, this makes things work and really tags have no place in simulation so should be correct going foward.