Rework UbooneBlobSource to operate on a "per view" mode.
Add snakemake script to run each view in parallel.

This uncovered a bug due to this line:

 gInterpreter->GenerateDictionary("vector<vector<int> >", "vector");

It causes ROOT to poop some code into the PWD with non-unique name.
That is not even multi-process safe.  No idea what mayhem MT would
cause.

That was added as the v<v<i>> dictionary does not exist.

Fix there is to make these dictionaries in a LinkDef.h.

And THAT uncovered some bugs in our waft tools support for rootcling
which this commit also fixes.