Remove -Werror.
This is causing too much headache due to different developers
catching/causing different warnings.  After developers stop using
ancient compilers, perhaps we can reinstate it.

Until then: SEE A WARNING FIX A WARNING.

Reference: #305.

This commit also includes the last batch of fixes prior to removal of
the flag.  Some of the fixed code is ancient and it is a mystery to me
why it never tickled -Werror before.