Make Pgraph deterministic.
Previously the topographical sort order is "random" based based on whatever
memory addresses located the nodes.

This commit adds an "instance" counter on each node which simply counts the
order of node construction.  This instance is then used to order the nodes in
the topographical sort.

In full jobs, the order of construction is determined by the order of the
component list from the Jsonnet config and Jsonnet is deterministic.  It is
possible to construct two Jsonnet configs that otherwise produce the same graph
but with different component list order.  Such jobs may see different Pgraph
execution order.