Make the build modal depending on git branch.
If "master" or a release branch then the build is "release mode" else it is
"development mode".

In release mode, disable any packages not ready for release (eg patrec) and make
the compiler more strict.  Fixes #305.

Currently, development mode (building in a feature branch) has no changes from
nominal.

This commit improves setting VERSION and adds support for providing the version
in a "version.txt" file when the source is not under git control.  Fixes #112.