Add a dirx() to IAnodeFace.
Partly addresses #373.

It returns +1 for a face that "looks" in the positive global-X
direction or -1 if it "looks" in the opposite direction.  Jumpered
AnodePlanes (icarus/sbnd) will have the same dirx for both faces,
wrapped AnodePlane faces will have opposite signs.

For now, the dirx value is deduced from the anode_x and respnose_x
given to AnodePlane, eg in the "volumes" configuration data structure.

I had thought the dirx can be calcualted from wires / pimpos (ie,
wire-cross-pitch give normal) but that approach gave -1 for all PDSP
faces.  I suspect either the (corrected) PDSP wires do not follow the
convention OR the convention is not so well defined and actually
induces this ambiguity.

Added a new WCT "app" to surface the "dirx" and other values of the
AnodePlane hierarchy.

wire-cell -A detector=uboone apps/test/anode-dumper.jsonnet
wire-cell -A detector=pdsp apps/test/anode-dumper.jsonnet