Possibly address #383.
The LMN resampling was recently merged after sitting a very long time in a
forgotten branch.  It is now applied to resample from FR (100ns) to ADC (500ns)
sampling.  Previously this was done in a way that was not quite precise as was
uncovered during the LMN development.

But, this initial application apparently chose the wrong normalization
intepretation applied as per the comments in the new spectrum_resize() function
in PlaneImpactResponse.  Because the new code does not expliciltly multiply by
the "rawresp_tick" we want to use the "corollary" case where the norm is the
value of the "old tick" (FR tick).

This change brings the SIM+SP peaks close to those from 0.29.3.

However, the SP signals are not identical.  To some extent, this is expected
because the prior resampling in the sim was not precise.  Some more
investigation at ADC level is needed.