Add intersect(line) and crop(segment) to BoundingBox.
The first finds intersection of an infinite line with a BB, the second
assumes a finite line segment.

This may be used in conjunction with new wpid->BB method of
IDetectorVolume to split up a long segment into subsegments for each
drift volume.

Add plane split and plane intersection functions to Point in support
of this.

Tests for these new functions in util/test/doctest_point.cxx

Partly addresses #393.