Fix #411 by making clustering methods into real components.
Important caveats:

- still need to bring wcp-porting-img config into compliance.

- an intermediate getClusteringMethod() function is still in use in order to
  make the configuration changes mininimal.

Config change FOR NOW is to change:

  name:"clustering_function_name"

to

  name:"ClusteringFunctionName:uniquetn"

and be sure to add ":uniquetn" in to assure all instances are unique.

Next step will be to configure each clustering method instance separate (like
normal components) and change MABC's "func_cfgs" to be a simple list of
type/names.