Introduce facade "flags" to restore "cluster_connected_dead" information.
The "live dead" clustering functino now sets a "live_dead" flag on live clusters
that are connected to dead clusters instead of placing these clusters in the
"cluster_connected_dead" set, which is now gone.  This flag (and any others) is
forwarded through "merge clusters" so a "fresh cluster" gets it if any of its
donor clusters had it.  The local set<Cluster*> that was passed between the
"extend" functions is gone and they instead check for this flag.

Some other minor cleanups to move data that was held directly by the Cluster
facade into node PCs (ident, t0).  Other items still remain.

Added a from() method to bundle copying of data "from" another cluster to this
cluster.  It copies any of these just added flags as well as default scope,
filter and transform info.  It is called from merge_clusters().  Various other
code contexts can be simplified to call from() instead of copy-pasting but those
are not yet cleaned up.