Fix cache fill order of operations.
The last fix introduced a somewhat subtle bug causing the m_nfkd pointer to be
nulled as part of its creation.  This was due to ScopedView::selections()
causing a cache check which was a cache miss and thus invalidate() called and
thus the just created m_nfkd pointer was destroyed.  This fix gets the
selections before making the m_nfkd pointer and all seems well again.