Extend UbooneClusterSource to remake all live slices from ROOT data.
This will produce fresh slices from ROOT data and these slices are in general a
superset of the union of all slices as slices from blob-free time slices will be
included.  The activity map for a fresh slice is made directly from ROOT data
and does not include any "bodge" for 2-plane views as done in the upstream
UbooneBlobSource's.  Activity in a dead channel/slice are still marked with the
conventional (0, 1e2) signal measure and unsertainty.

This also changes the API for add_ctpc() and add_dead_winds() helper functions
to be more straightforward taking a vector of ISlice instead of a vector of
IBlobSet, just to get a slice from each blob set.