Fix non-C++-17 code.
C++17 does not allow "structured bindings" to be used in "lambda capture list
restrictions".  C++20 is more lenient and apparently so is g++ with using C++17
standard.  However clang++ is not so lenient and this non-C++-17 code is caught as an error.

Also, fix one llm-coded compiler warning missed in a test from a while ago.